Revert "Branch acervo"

+++ include/acervo.h
@@ -7,7 +7,7 @@ using namespace std;
 class Acervo {
 public:
     //Construtor
-    Acervo(string autor, int anoPublicacao, string titulo, string genero, int codigo);
+    Acervo(string autor, int anoPublicacao, string titulo, string genero,float codigo);
     virtual ~Acervo();
 
     //getters
@@ -15,14 +15,14 @@ public:
     int getAnoPublicacao() const;
     string getTitulo() const;
     string getGenero() const;
-    int getCodigo() const;
+    float getCodigo() const;
 
     //setters
     /*string setAutor() const;
     int setAnoPublicacao() const;
     string setTitulo() const;
     string setGenero() const;
-    int setCodigo() const;*/
+    float setCodigo() const;*/
 
 
 private:
@@ -30,7 +30,7 @@ private:
     int anoPublicacao;
     string titulo;
     string genero;
-    int codigo;
+    float codigo;
 };
 
 #endif  // ACERVO_H
